+++ src/main/java/com/sprint/mission/discodeit/dto/UserDTO.java
@@ -22,6 +22,23 @@ public class UserDTO {
             byte[] profileImage,
             FileType fileType) {
 
+        public boolean isEmailValid(String email) {
+            return email.matches( "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+$");
+        }
+
+        public boolean isPasswordValid(String password) {
+            //길이는 8자리 이상, 영어 대소문자, 숫자, 특수문자 포함
+            return password.matches("^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%^&*])[A-Za-z\\d!@#$%^&*]{8,}$");
+        }
+
+        public CreateUserRequest {
+
+            if (!isPasswordValid(password) || !isEmailValid(email) || nickname.isBlank()) {
+                throw new IllegalArgumentException("Invalid user data.");
+            }
+
+        }
+
     }
 
     //user update를 위한 Request DTO
@@ -38,6 +55,23 @@ public class UserDTO {
             FileType fileType
     ) {
 
+        public boolean isEmailValid(String email) {
+            return email.matches( "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+$");
+        }
+
+        public boolean isPasswordValid(String password) {
+            //길이는 8자리 이상, 영어 대소문자, 숫자, 특수문자 포함
+            return password.matches("^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%^&*])[A-Za-z\\d!@#$%^&*]{8,}$");
+        }
+
+        public UpdateUserRequest {
+
+            if (!isPasswordValid(newPassword) || !isEmailValid(email) || nickname.isBlank()) {
+                throw new IllegalArgumentException("Invalid user data.");
+            }
+
+        }
+
     }
 
     //find user을 위한 DTO
@@ -52,5 +86,7 @@ public class UserDTO {
     ) {
 
     }
+
+
 
 }
