+++ src/main/java/com/sprint/mission/discodeit/entity/BinaryContent.java
@@ -9,6 +9,7 @@ import java.util.UUID;
 
 @Getter
 @Builder
+@ToString
 public class BinaryContent implements Serializable {
 
     private static final long serialVersionUID = 1L;
